requirements.txt
# AI Meeting Summarizer - Python Dependencies
# Minimal setup for Euron.one + Whisper

# ============ CORE REQUIREMENTS (REQUIRED) ============

# Backend Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5

# Audio Processing
sounddevice==0.4.6
soundfile==0.12.1

# AI & Transcription
openai-whisper==20231117
openai==1.12.0

# Database
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.1
requests==2.31.0
python-dateutil==2.8.2
deep-translator==1.11.4


# ============ OPTIONAL PACKAGES ============
# Uncomment only if you need these features

# PyAudio (alternative audio library - may have install issues on Windows)
# pyaudio==0.2.14

# Anthropic/Claude API support (not needed for Euron.one)
# anthropic==0.18.1

# LangChain/LangGraph (advanced AI workflows - not needed for basic usage)
# langchain==0.1.9
# langchain-openai==0.0.6
# langgraph==0.0.25

# CrewAI (multi-agent AI - not needed for basic usage)
# crewai==0.1.24

# Local LLM Support (only if using offline mode)
# llama-cpp-python==0.2.55
# transformers==4.37.2
# torch==2.2.0

# Pydantic (only needed for advanced validation)
# pydantic==2.6.1
# pydantic-settings==2.1.0

# Task Integration APIs (only if syncing to external services)
google-api-python-client==2.119.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
notion-client==2.2.1
jira==3.6.0

